+++ e2e/kubb.config.js
@@ -14,6 +14,7 @@ import { pluginZod } from '@kubb/plugin-zod'
 
 const schemas = [
   { name: 'test', path: './schemas/test.json' },
+  { name: 'notus', path: 'https://api.notus.team/openapi' },
   { name: 'Machines API', path: 'https://docs.machines.dev/spec/openapi3.json' },
   { name: 'discriminator', path: './schemas/discriminator.yaml' },
   { name: 'bunq.com', path: './schemas/bunq.com.json', strict: false },
@@ -46,6 +47,11 @@ const baseConfig = {
     clean: true,
   },
   plugins: [
+    pluginTs({
+      output: {
+        path: './types.ts',
+      },
+    }),
     pluginOas({
       generators: [],
       validate: false,