+++ docs/changelog.md
@@ -6,6 +6,8 @@ title: Changelog
 
 # 3.11.0
 - [`plugin-faker`](/plugins/plugin-faker): Min and max is not applied to the faker functions when only one of them is defined
+- [`core`](/plugins/core): uniqueBy for file.sources(isExportable and name)
+- [`plugin-ts`](/plugins/plugin-ts): Duplicated enums on typescript types
 
 # 3.10.13
 - [`plugin-zod`](/plugins/plugin-zod): query parameter objects are no longer optional if at least one parameter is defaulted